refactor : supabase.js를  객체로 수정
