+++ src/App.jsx
@@ -1,7 +1,7 @@
 import "./App.css";
 
 async function App() {
-  return <>{/* <Supabase /> */}</>;
+  return <>123</>;
 }
 
 export default App;
